Autenticando todas as rotas e  criando os validators de produto e delivery

+++ app/Controllers/Http/ProdutoController.js
@@ -7,6 +7,12 @@ const { is } = use("App/Helpers");
 class ProdutoController {
   async index({ response }) {
     try {
+      const user = await auth.getUser();
+      const verify = await is(user, "salesman");
+      if (!verify) {
+        const verify = await is(user, "admin");
+        if (!verify) return response.status(401).send("Usuário sem permissão");
+      }
       let produtos = await Database.raw(
         "SELECT id, nome, descricao, preco FROM produtos WHERE vendendo  = 1"
       );
@@ -22,17 +28,6 @@ class ProdutoController {
       const user = await auth.getUser();
       const verify = await is(user, "admin");
       if (!verify) return response.status(401).send("Usuário sem permissão");
-      /*console.log(ProdutoValidator.rules());
-
-      const validation = await validateAll(
-        request.all(),
-        await ProdutoValidator.rules(),
-        await ProdutoValidator.messages()
-      );
-
-      if (validation.fails()) {
-        return response.status(401).send({ message: validation.messages() });
-      }*/
       const data = request.only(["nome", "descricao", "preco"]);
       const produto = await Produto.create(data);
       return produto;
@@ -43,6 +38,12 @@ class ProdutoController {
 
   async show({ request, response }) {
     try {
+      const user = await auth.getUser();
+      const verify = await is(user, "salesman");
+      if (!verify) {
+        const verify = await is(user, "admin");
+        if (!verify) return response.status(401).send("Usuário sem permissão");
+      }
       let requisicao = request.only(["nome"]);
       let data = await Database.raw(
         `SELECT nome,descricao,preco FROM produtos WHERE nome LIKE '%${requisicao.nome}%'`
@@ -56,6 +57,9 @@ class ProdutoController {
 
   async update({ params, request, response }) {
     try {
+      const user = await auth.getUser();
+      const verify = await is(user, "admin");
+      if (!verify) return response.status(401).send("Usuário sem permissão");
       const data = request.all();
       let produto = await Produto.findByOrFail("id", params.id);
       let produtoAtualizado = {
@@ -73,6 +77,9 @@ class ProdutoController {
 
   async destroy({ params, response }) {
     try {
+      const user = await auth.getUser();
+      const verify = await is(user, "admin");
+      if (!verify) return response.status(401).send("Usuário sem permissão");
       await Produto.query().where("id", params.id).delete();
       return "Produto excluido com sucesso";
     } catch (error) {
